Move collection classes from the runtime module into their own module
